Update to merge with new code structure
